fix(landing): the highlight function has been replaced by highlight component.
