+++ apps/landing/components/Features/index.tsx
@@ -1,16 +1,16 @@
-import { highlight, useFormatMessage } from "@/hooks/intl";
+import { useFormatMessage } from "@/hooks/intl";
 import { FeatureProps } from "@/components/Features/types";
 import Feature from "@/components/Features/Feautre";
 import { LottieAnimate } from "@/components/Common/LottieAnimate";
+import { Highlight } from "@/components/Common/Highlight";
 
 export const Features: React.FC = () => {
   const intl = useFormatMessage();
   const features: Array<FeatureProps> = [
     {
       title: intl("home/features/feature/one-on-one/title"),
-      description: highlight(
-        intl,
-        "home/features/feature/one-on-one/description"
+      description: (
+        <Highlight id="home/features/feature/one-on-one/description" />
       ),
       image: (
         <LottieAnimate animation="oneOnOne" className="w-1/2 min-w-[350px]" />
@@ -19,7 +19,7 @@ export const Features: React.FC = () => {
     },
     {
       title: intl("home/features/feature/anytime/title"),
-      description: highlight(intl, "home/features/feature/anytime/description"),
+      description: <Highlight id="home/features/feature/anytime/description" />,
       image: (
         <LottieAnimate animation="anytime" className="w-1/2 min-w-[350px]" />
       ),
@@ -27,9 +27,8 @@ export const Features: React.FC = () => {
     },
     {
       title: intl("home/features/feature/novelity/title"),
-      description: highlight(
-        intl,
-        "home/features/feature/novelity/description"
+      description: (
+        <Highlight id="home/features/feature/novelity/description" />
       ),
       image: (
         <LottieAnimate animation="novelity" className="w-1/2 min-w-[350px]" />
@@ -38,9 +37,8 @@ export const Features: React.FC = () => {
     },
     {
       title: intl("home/features/feature/time-is-yours/title"),
-      description: highlight(
-        intl,
-        "home/features/feature/time-is-yours/description"
+      description: (
+        <Highlight id="home/features/feature/time-is-yours/description" />
       ),
       image: (
         <LottieAnimate
@@ -52,9 +50,8 @@ export const Features: React.FC = () => {
     },
     {
       title: intl("home/features/feature/all-in-english/title"),
-      description: highlight(
-        intl,
-        "home/features/feature/all-in-english/description"
+      description: (
+        <Highlight id="home/features/feature/all-in-english/description" />
       ),
       image: (
         <LottieAnimate
@@ -66,9 +63,8 @@ export const Features: React.FC = () => {
     },
     {
       title: intl("home/features/feature/notifications/title"),
-      description: highlight(
-        intl,
-        "home/features/feature/notifications/description"
+      description: (
+        <Highlight id="home/features/feature/notifications/description" />
       ),
       image: (
         <LottieAnimate
