+++ apps/landing/components/Tutors/index.tsx
@@ -4,14 +4,12 @@ import Content from "@/components/Tutors/Content";
 import { Typography } from "@litespace/ui/Typography";
 import React from "react";
 import InViewTrack from "@/components/Common/InViewTrack";
-import { highlight, useFormatMessage } from "@/hooks/intl";
 import { ITutor } from "@litespace/types";
+import { Highlight } from "@/components/Common/Highlight";
 
 export const Tutors: React.FC<{
   tutors: ITutor.FindOnboardedTutorsApiResponse;
 }> = ({ tutors }) => {
-  const intl = useFormatMessage();
-
   return (
     <div className="bg-natural-0 max-w-screen">
       <InViewTrack event="view_item_list" label="tutors" action="scroll" />
@@ -21,13 +19,13 @@ export const Tutors: React.FC<{
             tag="h3"
             className="text-subtitle-1 md:text-h4 lg:text-h3 font-bold text-natural-950"
           >
-            {highlight(intl, "home/tutors/title")}
+            <Highlight id="home/tutors/title" />
           </Typography>
           <Typography
             tag="h6"
             className="text-body md:text-subtitle-2 font-medium md:font-semibold  text-natural-700"
           >
-            {highlight(intl, "home/tutors/description")}
+            <Highlight id="home/tutors/description" />
           </Typography>
         </div>
